Linux: Remove start_disabled from PerfBuilder.
Upcoming additions will allow redirection of a perf event's output to
another. Such preparation work needs to be done before enabling the
event. Having logic in the constructor to enable the perf event means
that the constructor will be responsible for such preparation work as
well, which complicates the arguments.

As such, move the responsibility of enabling an event to the caller. We
don't have any use case that does not use start_disabled anyway.